create jobpost 3/6/2025

+++ src/main/java/com/example/swp391_d01_g3/repository/IJobPostRepository.java
@@ -6,5 +6,5 @@ import org.springframework.stereotype.Repository;
 
 
 @Repository
-public interface IJobPostRepository extends JpaRepository<JobPost,Long> {
+public interface IJobPostRepository extends JpaRepository<JobPost,Integer> {
 }
